Move processing to authenticated and unauthenticated fragments
