+++ app/src/main/res/layout/fragment_authenticated.xml
@@ -42,7 +42,7 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:text="@string/refresh_button"
-                android:onClick="@{() -> model.refreshToken()}"/>
+                android:onClick="@{() -> model.refreshAccessToken()}"/>
 
             <Button
                 android:id="@+id/logoutButton"
